UAS_15773

+++ gallery.php
@@ -3,7 +3,7 @@
         <div class="col-md-6">
             <!-- Button tambah data -->
             <button type="button" class="btn btn-secondary mb-2" data-bs-toggle="modal" data-bs-target="#modalTambah">
-             <i class="bi bi-plus-lg"></i> Tambah Article
+             <i class="bi bi-plus-lg"></i> Tambah Gambar
             </button>
         </div>
         <div class="col-md-6">
@@ -21,10 +21,9 @@
                 <thead class="table-dark">
                     <tr>
                         <th>No</th>
-                        <th class="w-25">Judul</th>
-                        <th class="w-50">Isi</th>
-                        <th class="w-50">Gambar</th>
-                        <th class="w-25">Aksi</th>
+                        <th class="w-50">Deskripsi</th>
+                        <th class="w-60">Gambar</th>
+                        <th> Aksi</th>
                     </tr>   
                 </thead>
                 <tbody id="result">
@@ -32,27 +31,23 @@
                 </tbody>
             </table>
         </div>
-        <!-- Modal Tambah Article -->
+        <!-- Modal Tambah Gambar -->
         <div class="modal fade" id="modalTambah" tabindex="-1" aria-labelledby="modalTambahLabel" aria-hidden="true">
             <div class="modal-dialog">
                 <div class="modal-content">
                     <div class="modal-header">
-                        <h1 class="modal-title fs-5" id="modalTambahLabel">Modal title</h1>
+                        <h1 class="modal-title fs-5" id="modalTambahLabel">Tambah Gambar</h1>
                         <button type="button" class="btn-close" data-bs-dismiss="modal" aria-label="Close"></button>
                     </div>
                     <form method="post" action="" enctype="multipart/form-data">
                     <div class="modal-body">
                         <div class="mb-3">
-							<label for="judul" class="form-label">Judul</label>
-                            <input type="text" class="form-control" name="judul" placeholder="Tuliskan Judul Artikel" required>
-                        </div>
-                        <div class="mb-3">
-                            <label for="isi">Isi</label>
-                            <textarea class="form-control" placeholder="Tuliskan Isi Artikel" name="isi" required></textarea>
+							<label for="deskripsi" class="form-label">Deskripsi</label>
+                            <input type="text" class="form-control" name="deskripsi" placeholder="Tuliskan Deskripsi Gambar" required>
                         </div>
                         <div class="mb-3">
                             <label for="gambar" class="form-label">Gambar</label>
-                            <input type="file" class="form-control" name="gambar">
+                            <input type="file" class="form-control" name="gambar" required>
                         </div>
                     </div>
                     <div class="modal-footer">
@@ -69,7 +64,7 @@
 <script>
     function loadData(keyword = '') {
         $.ajax({
-            url: "article_search.php",
+            url: "gallery_search.php",
             type: "POST",
             data: {
                 keyword: keyword
@@ -100,8 +95,7 @@ include "upload_photo.php";
 
 //jika tombol simpan diklik
 if (isset($_POST['simpan'])) {
-    $judul = $_POST['judul'];
-    $isi = $_POST['isi'];
+    $deskripsi = $_POST['deskripsi'];
     $tanggal = date("Y-m-d H:i:s");
     $username = $_SESSION['username'];
     $gambar = '';
@@ -118,16 +112,16 @@ if (isset($_POST['simpan'])) {
             //jika true maka message berisi nama file gambar
             $gambar = $cek_upload['message'];
         } else {
-            //jika true maka message berisi pesan error, tampilkan dalam alert
+            //jika false maka message berisi pesan error, tampilkan dalam alert
             echo "<script>
                 alert('" . $cek_upload['message'] . "');
-                document.location='admin.php?page=article';
+                document.location='admin.php?page=gallery';
             </script>";
             die;
         }
     }
 
-    		//cek apakah ada id yang dikirimkan dari form
+    	//cek apakah ada id yang dikirimkan dari form
     if (isset($_POST['id'])) {
         //jika ada id, lakukan update data dengan id tersebut
         $id = $_POST['id'];
@@ -137,38 +131,39 @@ if (isset($_POST['simpan'])) {
             $gambar = $_POST['gambar_lama'];
         } else {
             //jika ganti gambar, hapus gambar lama
-            unlink("img/" . $_POST['gambar_lama']);
+            if (file_exists("img/" . $_POST['gambar_lama'])) {
+                unlink("img/" . $_POST['gambar_lama']);
+            }
         }
 
-        $stmt = $conn->prepare("UPDATE article 
+        $stmt = $conn->prepare("UPDATE gallery 
                                 SET 
-                                judul =?,
-                                isi =?,
+                                deskripsi = ?,
                                 gambar = ?,
                                 tanggal = ?,
                                 username = ?
                                 WHERE id = ?");
 
-        $stmt->bind_param("sssssi", $judul, $isi, $gambar, $tanggal, $username, $id);
+        $stmt->bind_param("ssssi", $deskripsi, $gambar, $tanggal, $username, $id);
         $simpan = $stmt->execute();
     } else {
-		    //jika tidak ada id, lakukan insert data baru
-        $stmt = $conn->prepare("INSERT INTO article (judul,isi,gambar,tanggal,username)
-                                VALUES (?,?,?,?,?)");
+	    //jika tidak ada id, lakukan insert data baru
+        $stmt = $conn->prepare("INSERT INTO gallery (deskripsi, gambar, tanggal, username)
+                                VALUES (?, ?, ?, ?)");
 
-        $stmt->bind_param("sssss", $judul, $isi, $gambar, $tanggal, $username);
+        $stmt->bind_param("ssss", $deskripsi, $gambar, $tanggal, $username);
         $simpan = $stmt->execute();
     }
 
     if ($simpan) {
         echo "<script>
             alert('Simpan data sukses');
-            document.location='admin.php?page=article';
+            document.location='admin.php?page=gallery';
         </script>";
     } else {
         echo "<script>
             alert('Simpan data gagal');
-            document.location='admin.php?page=article';
+            document.location='admin.php?page=gallery';
         </script>";
     }
 
@@ -183,10 +178,12 @@ if (isset($_POST['hapus'])) {
 
     if ($gambar != '') {
         //hapus file gambar dari folder /img
-        unlink("img/" . $gambar);
+        if (file_exists("img/" . $gambar)) {
+            unlink("img/" . $gambar);
+        }
     }
 
-    $stmt = $conn->prepare("DELETE FROM articles WHERE id =?");
+    $stmt = $conn->prepare("DELETE FROM gallery WHERE id = ?");
 
     $stmt->bind_param("i", $id);
     $hapus = $stmt->execute();
@@ -194,12 +191,12 @@ if (isset($_POST['hapus'])) {
     if ($hapus) {
         echo "<script>
             alert('Hapus data sukses');
-            document.location='admin.php?page=article';
+            document.location='admin.php?page=gallery';
         </script>";
     } else {
         echo "<script>
             alert('Hapus data gagal');
-            document.location='admin.php?page=article';
+            document.location='admin.php?page=gallery';
         </script>";
     }
 
@@ -207,4 +204,4 @@ if (isset($_POST['hapus'])) {
     $conn->close();
 }
 
-?>  
+?>
